requirements.txt
BeautifulSoup==3.2.1
Django==1.5.5
MySQL-python==1.2.5
Pillow==2.5.0
South==1.0
beautifulsoup4==4.3.2
django-debug-toolbar==1.2.1
django-filebrowser==3.5.6
django-grappelli==2.4.10
django-qrcode==0.3
django-tastypie==0.12.0
docutils==0.11
gnureadline==6.3.3
ipython==2.1.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.8.0
sqlparse==0.1.11
wsgiref==0.1.2
